K-Means w/ Neighbors Version-2
This version of the K-Means algorithm that includes the neighboring pixels only takes into account the neighboring pixels, if the current pixel being worked on is not a part of the boundary of the image. Basically, if the pixel is either a corner or any of the surrounding edges of the image, then they do not include the neighboring pixels. However, every other pixel takes into account their neighboring pixels.